requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Caching==2.0.2
Flask-WTF==1.2.1
Flask-Mail==0.9.1
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
plotly==5.18.0
dash==2.14.2
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
msal==1.24.0
python-dotenv==1.0.0
Jinja2==3.1.2
Werkzeug==3.0.1
SQLAlchemy==2.0.23
alembic==1.12.1
WTForms==3.1.1
email-validator==2.1.0
reportlab==4.0.4
weasyprint==61.0
python-dateutil==2.8.2
pytz==2023.3